% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.proteoNorm.R
\name{print.proteoNorm}
\alias{print.proteoNorm}
\title{Print Method for proteoNorm Objects}
\usage{
\method{print}{proteoNorm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{proteoNorm}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Custom print method for objects of class \code{proteoNorm}.
}
